# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "API documentation for Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Blog/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/blog-content": {
            "get": {
                "tags": [
                    "Blog Content"
                ],
                "summary": "all Posts and all Comments and all available Reactos",
                "description": "Get all posts and all comments connected to every post and reactos connected to every post and comment and all available reactos ",
                "operationId": "b440fb7fd4f5b2433893997bc177ffff",
                "responses": {
                    "200": {
                        "description": "all Posts and all Comments and all available Reactos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All Posts with the Comments and Reactos of every Post and Comment Retrieved Successfully, and all available Reactos"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/blog-content"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": "http://localhost/Blog/public/api/blog-content?page=1",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "1"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "reactos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Unauthorized Action"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/blog-content/search-user": {
            "get": {
                "tags": [
                    "Blog Content"
                ],
                "summary": "Retrieve Users resulted from Search.",
                "description": "Retrieve all available Users resulted from the search.",
                "operationId": "0fda3c3a888b1d07b56a52997fa5e743",
                "parameters": [
                    {
                        "name": "nameKey",
                        "in": "query",
                        "description": "Search keyword",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "aka"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully(may be empty).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "All available Users resulted from the search Retrieved Successfully."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "There is no available Users resulted from the search."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/blog-content/posts-user/{id}": {
            "get": {
                "tags": [
                    "Blog Content"
                ],
                "summary": "all Posts and all Comments and all of a User and all available Reactos",
                "description": "Get all posts and all comments connected to every post, and reactos connected to every post and comment of a User and all available reactos.",
                "operationId": "38acc8cb27afa409da363971c8a32f24",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all posts and all comments connected to every post, and reactos connected to every post and comment of a User and all available reactos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "a User All Posts with the Comments and Reactos of every Post and Comment Retrieved Successfully, and all available Reactos"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/posts?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/blog-content/posts-user/{id}"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": "http://localhost/Blog/public/api//blog-content/posts-user/{id}?page=1",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "1"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "reactos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lang": {
                        "type": "string",
                        "example": "user lang"
                    },
                    "body": {
                        "type": "string",
                        "example": "post content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    },
                    "reactos": {
                        "description": "Loaded only if relation reactos is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReactResource"
                        }
                    }
                },
                "type": "object"
            },
            "ReactResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "tag name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "posts": {
                        "description": "Loaded only if relation posts is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostResource"
                        }
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "user name"
                    },
                    "email": {
                        "type": "string",
                        "example": "name@test.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "user phone"
                    },
                    "lang": {
                        "type": "string",
                        "example": "user lang"
                    },
                    "favicon": {
                        "type": "string",
                        "example": "user favicon"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "posts": {
                        "description": "Loaded only if relation posts is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostResource"
                        }
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    },
                    "roles": {
                        "description": "Loaded only if relation roles is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResource"
                        }
                    },
                    "permissions": {
                        "description": "Loaded only if relation roles is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResource"
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}